# docs/api/openapi.yaml
components:
  schemas:
...
    AlertHistoryResponse:
      additionalProperties: false
      description: One customer-visible notification.
      properties:
        action_page:
          anyOf:
          - type: string
          - type: 'null'
          title: Action Page
        created_at:
          title: Created At
          type: string
        domain:
          title: Domain
          type: string
        event_type:
          title: Event Type
          type: string
        id:
          title: Id
          type: string
        is_read:
          title: Is Read
          type: boolean
        message:
          title: Message
          type: string
        title:
          title: Title
          type: string
      required:
      - id
      - domain
      - event_type
      - title
      - message
      - created_at
      - is_read
      title: AlertHistoryResponse
      type: object
    AlertPreferenceResponse:
      additionalProperties: false
      description: Notification preferences for one connected domain.
      properties:
        domain:
          title: Domain
          type: string
        email_enabled:
          title: Email Enabled
          type: boolean
        notify_domain_shield:
          title: Notify Domain Shield
          type: boolean
        notify_phishing:
          title: Notify Phishing
          type: boolean
        quiet_hours_enabled:
          title: Quiet Hours Enabled
          type: boolean
        quiet_hours_end:
          title: Quiet Hours End
          type: string
        quiet_hours_start:
          title: Quiet Hours Start
          type: string
        timezone:
          title: Timezone
          type: string
      required:
      - domain
      - email_enabled
      - notify_phishing
      - notify_domain_shield
      - quiet_hours_enabled
      - quiet_hours_start
      - quiet_hours_end
      - timezone
      title: AlertPreferenceResponse
      type: object
    AlertPreferenceUpdate:
      properties:
        email_enabled:
          title: Email Enabled
          type: boolean
        notify_domain_shield:
          title: Notify Domain Shield
          type: boolean
        notify_phishing:
          title: Notify Phishing
          type: boolean
        quiet_hours_enabled:
          title: Quiet Hours Enabled
          type: boolean
        quiet_hours_end:
          default: 07:00
          pattern: ^(?:[01]\d|2[0-3]):[0-5]\d$
          title: Quiet Hours End
          type: string
        quiet_hours_start:
          default: '22:00'
          pattern: ^(?:[01]\d|2[0-3]):[0-5]\d$
          title: Quiet Hours Start
          type: string
        timezone:
          default: Europe/Paris
          maxLength: 64
          minLength: 1
          title: Timezone
          type: string
      required:
      - email_enabled
      - notify_phishing
      - notify_domain_shield
      - quiet_hours_enabled
      title: AlertPreferenceUpdate
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    SecurityRuleCreate:
      properties:
        pattern:
          maxLength: 254
          minLength: 3
          pattern: ^(?:[^@\s]+@)?(?:[A-Za-z0-9](?:[A-Za-z0-9-]{0,61}[A-Za-z0-9])?\.)+[A-Za-z]{2,63}$
          title: Pattern
          type: string
        rule_type:
          pattern: ^(whitelist|blocklist)$
          title: Rule Type
          type: string
      required:
      - rule_type
      - pattern
      title: SecurityRuleCreate
      type: object
    SecurityRuleResponse:
      additionalProperties: false
      description: Workspace sender or domain rule.
      properties:
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
        domain:
          title: Domain
          type: string
        id:
          title: Id
          type: string
        pattern:
          title: Pattern
          type: string
        rule_type:
          enum:
          - whitelist
          - blocklist
          title: Rule Type
          type: string
      required:
      - id
      - domain
      - rule_type
      - pattern
      title: SecurityRuleResponse
      type: object
...
    StatusResponse:
      additionalProperties: false
      description: Stable acknowledgement for a completed command.
      properties:
        status:
          title: Status
          type: string
      required:
      - status
      title: StatusResponse
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: Generated from the deployed FastAPI routes and Pydantic models. Authenticated customer
    operations use Better Auth sessions or bearer tokens; internal and email-gateway routes require their
    documented service credentials.
  summary: Data, application, and integration contracts for Sicurre.
  title: Sicurre Platform API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/alerts/history:
    get:
      operationId: list_alert_history_v1_alerts_history_get
      parameters:
      - in: query
        name: domain
        required: true
        schema:
          title: Domain
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AlertHistoryResponse'
                title: Response List Alert History V1 Alerts History Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List Alert History
      tags:
      - app-ui-flows
  /v1/alerts/history/read:
    post:
      operationId: mark_domain_alerts_read_v1_alerts_history_read_post
      parameters:
      - in: query
        name: domain
        required: true
        schema:
          title: Domain
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Mark Domain Alerts Read
      tags:
      - app-ui-flows
  /v1/alerts/history/{id}/dismiss:
    post:
      operationId: dismiss_alert_v1_alerts_history__id__dismiss_post
      parameters:
      - in: path
        name: id
        required: true
        schema:
          title: Id
          type: string
      - in: query
        name: domain
        required: true
        schema:
          title: Domain
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Dismiss Alert
      tags:
      - app-ui-flows
  /v1/alerts/history/{id}/read:
    post:
      description: Mark one owned, domain-scoped notification as read for this member.
      operationId: mark_alert_read_v1_alerts_history__id__read_post
      parameters:
      - in: path
        name: id
        required: true
        schema:
          title: Id
          type: string
      - in: query
        name: domain
        required: true
        schema:
          title: Domain
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Mark Alert Read
      tags:
      - app-ui-flows
  /v1/alerts/preferences:
    get:
      operationId: get_alert_preferences_v1_alerts_preferences_get
      parameters:
      - in: query
        name: domain
        required: true
        schema:
          title: Domain
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertPreferenceResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Alert Preferences
      tags:
      - app-ui-flows
    put:
      operationId: update_alert_preferences_v1_alerts_preferences_put
      parameters:
      - in: query
        name: domain
        required: true
        schema:
          title: Domain
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertPreferenceUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Update Alert Preferences
      tags:
      - app-ui-flows
  /v1/alerts/rules:
    get:
      operationId: list_security_rules_v1_alerts_rules_get
      parameters:
      - in: query
        name: domain
        required: true
        schema:
          title: Domain
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SecurityRuleResponse'
                title: Response List Security Rules V1 Alerts Rules Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List Security Rules
      tags:
      - app-ui-flows
    post:
      operationId: create_security_rule_v1_alerts_rules_post
      parameters:
      - in: query
        name: domain
        required: true
        schema:
          title: Domain
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityRuleCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityRuleResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Create Security Rule
      tags:
      - app-ui-flows
  /v1/alerts/rules/{id}:
    delete:
      operationId: delete_security_rule_v1_alerts_rules__id__delete
      parameters:
      - in: path
        name: id
        required: true
        schema:
          title: Id
          type: string
      - in: query
        name: domain
        required: true
        schema:
          title: Domain
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Delete Security Rule
      tags:
      - app-ui-flows
...
servers:
- description: Current Sicurre deployment
  url: /
...